requirements.txt
streamlit>=1.36
plotly>=5.22
duckdb>=1.0
polars>=1.5
requests>=2.32
python-dotenv>=1.0
pydantic>=2.7
scikit-learn>=1.5

ruff>=0.5
black>=24.4
mypy>=1.10
pytest>=8.2
pre-commit>=3.7
